refactor: move deserializer out of sub dir
